Move all plugins to ./plugin
Define them dynamically each in its own file with the Plugin objects.
Packer will then parse these plugin objects
