+++ init.lua
@@ -84,8 +84,3 @@ vim.g.loaded_netrw = 1
 vim.g.loaded_netrwPlugin = 1
 vim.g.loaded_netrwSettings = 1
 vim.g.loaded_netrwFileHandlers = 1
-
------------------------------------------------------------------------ PLUGINS
--- NOTE: plugins are defined in lua/plugins/, call the setup
--- function that configures them.
-require("plugins").setup()
